@@ -1,11 +1,10 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 
 import PackageDescription
 
 let settings: [SwiftSetting] = [
     .unsafeFlags(["-warnings-as-errors"]),
-    .enableExperimentalFeature("ExistentialAny"),
-    .enableExperimentalFeature("StrictConcurrency=targeted")
+    .enableExperimentalFeature("ExistentialAny")
 ]
 
 let package = Package(
@@ -34,5 +33,6 @@ let package = Package(
                 .product(name: "ConcurrencyExtras", package: "swift-concurrency-extras")
             ],
             swiftSettings: settings)
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )
