@@ -5,7 +5,7 @@ import Foundation
 /// Discussion: Adding this protocol is mean to extract the ability of formatting a `Date` object out of `DateFormatter` and shadow other abilities that a
 /// `DateFormatter`. For example, instead of returning `DateFormatter` from a dateFormatterFactory, returning the shadow - `DateFormatting`
 /// which end users to alter the formatter while using.
-@objc public protocol DateFormatting {
+@objc public protocol DateFormatting: Sendable {
     func localisedString(from date: Date) -> String
 }
 
